Test fetchArticle php 8.1 attributes
